We now always use HTTPS for the index page and the pages for creating, editing and deleting integration profiles on the website. Other pages, including the ones used to implement the integration (CaseFeed/, WebHook/ and URLTrigger/) continue to support both HTTP and HTTPS.
We also fixed a small but damaging bug where the FogBugz token saved for a user wasn't cleared when the login credentials (FogBugz base URL and user email) were changed, making FogTracker effectively still using the old user through the old token even if a new user has been specified. We also now don't require the FogBugz password be set when editing an existing integration profile, and if not set, it's simply not changed.